+++ panel/main.go
@@ -57,6 +57,7 @@ func main() {
 	st.prune(24*time.Hour, 20000)
 	st.pruneHourly(time.Duration(cfg.RetentionDays) * 24 * time.Hour)
 	st.pruneTargetTraffic(time.Duration(cfg.RetentionDays) * 24 * time.Hour)
+	st.pruneOutboundTraffic(time.Duration(cfg.RetentionDays) * 24 * time.Hour)
 	go func() {
 		t := time.NewTicker(time.Hour)
 		defer t.Stop()
@@ -68,6 +69,7 @@ func main() {
 				st.prune(24*time.Hour, 20000)
 				st.pruneHourly(time.Duration(cfg.RetentionDays) * 24 * time.Hour)
 				st.pruneTargetTraffic(time.Duration(cfg.RetentionDays) * 24 * time.Hour)
+				st.pruneOutboundTraffic(time.Duration(cfg.RetentionDays) * 24 * time.Hour)
 			}
 		}
 	}()